- Extracted all specific pokemon data sync grids from `GridMaps/PokemonData/SyncGridData.js` and isolated it for easier singular use. - Created a reusable `GridMap` component that can now be shared by all pokemon (prevents from having a single component per each pokemon)
